修改统计接口，使用实体类传输

+++ app/src/main/java/net/onest/time/api/vo/statistic/StatisticVo.java
@@ -10,7 +10,7 @@ public class StatisticVo {
     private Integer tomatoTimes;                                // 总番茄钟数（总专注次数）
     private Long tomatoDuration;                                // 总番茄钟时长（总专注时长）
     private Map<Long, DayTomatoStatistic> dayTomatoMap;         // 某日番茄钟统计
-    private Integer avgTomatoDuration;                          // 每天平均专注时长
+    private Long avgTomatoDuration;                          // 每天平均专注时长
     private Integer avgTomatoTimes;                             // 每天平均专注次数
     private Boolean cached;                                     // 是否是Redis缓存
     private Integer tomatoDays;                                 // 一共使用的天数
@@ -42,11 +42,11 @@ public class StatisticVo {
         this.dayTomatoMap = dayTomatoMap;
     }
 
-    public Integer getAvgTomatoDuration() {
+    public Long getAvgTomatoDuration() {
         return avgTomatoDuration;
     }
 
-    public void setAvgTomatoDuration(Integer avgTomatoDuration) {
+    public void setAvgTomatoDuration(Long avgTomatoDuration) {
         this.avgTomatoDuration = avgTomatoDuration;
     }
